requirements.txt
future==0.15.2
Pygments==2.1.3
pyqode.core==2.10.0
pyqode.qt==2.10.0
PyQt5==5.7
python-decouple==3.0
sip==4.18.1
SQLAlchemy==1.0.14
